milestone-2/requirements.txt
streamlit
google-generativeai
pillow
pandas
streamlit-mic-recorder
streamlit-TTS
gTTS
